[#2] chore: PostRepositoryError 파일을 별도로 분리
